Adds a filter to count bad filenames.
We want to save all generated data, but we also have best practices.
This code adds a filename linter to check best practices for the
filenames of saved data, and it also adds a prometheus counter to count
how many files we find with filenames that don't follow our best
practices.  Ideally this counter will always be zero.