dirs in metrics are now datatypes
the filename types are in their own package. This became necessary
because in adding the second string argument (for the purpose of
monitoring) it became clear that having 2 strings right in a row was
just asking for trouble.  Now the datatype is a string, but the filename
should always be either a filename.System or filename.Internal